release v2.2.0

+++ functions/front/carts.php
@@ -10,7 +10,13 @@
 
         if (isset($value) && $value !== FALSE) {
             $cart_item_data[PRINT_APP_CUSTOMIZATION_KEY] = $value;
-            General\delete_customization_data($product_id);
+            
+            // Don't delete during cart simulation or validation checks
+            $is_simulation = isset($_GET['wc-ajax']) && $_GET['wc-ajax'] === 'ppc-simulate-cart';
+            
+            if (!$is_simulation) {
+                General\delete_customization_data($product_id);
+            }
         }
 
         return $cart_item_data;
